♻️ refactor: separate message slice and aiChat slice (#4359)
* ♻️ refactor: separate message slice and ai-chat slice

* 🎨 chore: clean code

* ♻️ refactor: refactor the initialState

* ♻️ refactor: refactor the initialState

* ✅ test: add more tests
