+++ src/store/chat/initialState.ts
@@ -4,15 +4,18 @@ import { ChatPortalState, initialChatPortalState } from './slices/portal/initial
 import { ChatMessageState, initialMessageState } from './slices/message/initialState';
 import { ChatShareState, initialShareState } from './slices/share/initialState';
 import { ChatTopicState, initialTopicState } from './slices/topic/initialState';
+import { ChatAIChatState, initialAiChatState } from './slices/aiChat/initialState';
 
 export type ChatStoreState = ChatTopicState &
   ChatMessageState &
+  ChatAIChatState &
   ChatToolState &
   ChatShareState &
   ChatPortalState;
 
 export const initialState: ChatStoreState = {
   ...initialMessageState,
+  ...initialAiChatState,
   ...initialTopicState,
   ...initialToolState,
   ...initialShareState,
