+++ src/store/chat/slices/message/initialState.ts
@@ -1,18 +1,12 @@
 import { ChatMessage } from '@/types/message';
 
 export interface ChatMessageState {
-  abortController?: AbortController;
   /**
    * @title 当前活动的会话
    * @description 当前正在编辑或查看的会话
    */
   activeId: string;
-  /**
-   * is the AI message is generating
-   */
-  chatLoadingIds: string[];
-  inputFiles: File[];
-  inputMessage: string;
+
   isCreatingMessage: boolean;
   /**
    * is the message is editing
@@ -23,32 +17,17 @@ export interface ChatMessageState {
    */
   messageLoadingIds: string[];
   /**
-   * is the message is in RAG flow
-   */
-  messageRAGLoadingIds: string[];
-  /**
    * whether messages have fetched
    */
   messagesInit: boolean;
   messagesMap: Record<string, ChatMessage[]>;
-  pluginApiLoadingIds: string[];
-  /**
-   * the tool calling stream ids
-   */
-  toolCallingStreamIds: Record<string, boolean[]>;
 }
 
 export const initialMessageState: ChatMessageState = {
   activeId: 'inbox',
-  chatLoadingIds: [],
-  inputFiles: [],
-  inputMessage: '',
   isCreatingMessage: false,
   messageEditingIds: [],
   messageLoadingIds: [],
-  messageRAGLoadingIds: [],
   messagesInit: false,
   messagesMap: {},
-  pluginApiLoadingIds: [],
-  toolCallingStreamIds: {},
 };
