+++ src/store/chat/slices/share/action.test.ts
@@ -3,7 +3,7 @@ import { act, renderHook } from '@testing-library/react';
 import { DEFAULT_USER_AVATAR_URL } from '@/const/meta';
 import { shareService } from '@/services/share';
 import { useChatStore } from '@/store/chat';
-import { messageMapKey } from '@/store/chat/slices/message/utils';
+import { messageMapKey } from '@/store/chat/utils/messageMapKey';
 import { ChatMessage } from '@/types/message';
 
 describe('shareSlice actions', () => {
